+++ opal/porous/application.rb
@@ -4,9 +4,11 @@ module Porous
   class Application
     include Porous::Component
 
+    inject Porous::Router, as: :router
+
     def render
       body class: 'bg-gray-50 dark:bg-gray-900' do
-        component Porous::Router
+        component router
       end
     end
   end