+++ opal/porous/router.rb
@@ -1,4 +1,7 @@
+# frozen_string_literal: true
+
 module Porous
+  # rubocop:disable Metrics/ClassLength
   class Router
     include Porous::Component
 
@@ -10,7 +13,7 @@ module Porous
       # Object.descendants.each { |klass| puts klass if klass.is_a?(Class) }
       Object.descendants.select { |c| c.is_a?(Class) && c.included_modules.include?(Porous::Page) }.each do |klass|
         # puts "Included in: #{klass}"
-        next if klass.to_s.start_with? '#<Class:' # skip singleton classes
+        # next if klass.to_s.start_with? '#<Class:' # skip singleton classes
 
         @routes.combine klass.new.route!
       end
@@ -22,26 +25,39 @@ module Porous
       add_listeners
     end
 
-    def self.included(base)
-      base.extend(Porous::Component::ClassMethods)
-      Component.module_eval do
-        unless respond_to?(:__a)
-          alias_method :__a, :a
-          define_method(:a) do |params = {}, &block|
-            params = { onclick: ->(e) {
-                                  router.go_to(e.target.pathname) unless params[:target] == "_blank"
-                                } }.merge(params)
-            __a(params, &block)
+    # Handle anchor tags with router (requires router to be injected)
+    Component.module_eval do
+      unless respond_to?(:__a)
+        alias_method :__a, :a
+        define_method(:a) do |params = {}, &block|
+          if params['href'].include?('//') || params['target'] == '_blank'
+            # Retain behaviour
+          else
+            href = params['href']
+            params[:onclick] = lambda { |e|
+              e.prevent
+              raise Error, 'No router to handle navigation. Did you `inject Porous::Router`' unless router
+
+              router.go_to(href)
+            }
           end
+          __a(params, &block)
         end
       end
     end
 
     def add_listeners
-      if Browser::History.supported?
-        # TODO: Fix routing (navigating)
-        # Browser.on_pop_state { find_route; parse_url_params; render! }
-        # Browser.hash_change { find_route; parse_url_params; render! }
+      return unless Browser::History.supported?
+
+      $window.on(:popstate) do
+        find_route
+        parse_url_params
+        render!
+      end
+      $window.on(:hashchange) do
+        find_route
+        parse_url_params
+        render!
       end
     end
 
@@ -72,13 +88,13 @@ module Porous
     def call_on_enter_callback(route)
       return unless route[:on_enter]
 
-      if route[:on_enter].respond_to?(:call)
-        route[:on_enter].call
-      end
+      return unless route[:on_enter].respond_to?(:call)
+
+      route[:on_enter].call
     end
 
     def go_to(path)
-      Browser.push_state(path)
+      $window.history.push path
       find_route
       parse_url_params
       render!
@@ -86,15 +102,17 @@ module Porous
     end
 
     def parse_url_params
-      @params = compotent_url_params
-      query[1..-1].split('&').each do |param|
+      @params = component_url_params
+      return if query.empty?
+
+      query[1..].split('&').each do |param|
         key, value = param.split('=')
         @params[Browser.decode_uri_component(key)] = Browser.decode_uri_component(value)
-      end unless query.empty?
+      end
     end
 
-    def compotent_url_params
-      Hash[@route[:params].zip(path.match(@route[:regex])[1..-1])]
+    def component_url_params
+      @route[:params].zip(path.match(@route[:regex])[1..]).to_h
     end
 
     def url_for(name, params = nil)
@@ -112,11 +130,11 @@ module Porous
     end
 
     def query
-      $document.location.query
+      $window.location.query
     end
 
     def path
-      $document.location.path
+      $window.location.path
     end
 
     def current_url?(name)
@@ -125,10 +143,11 @@ module Porous
 
     def url_with_params(route, params)
       path = route[:path]
-      params.each do |key, value|
-        path = path.gsub(":#{key}", "#{value}")
-      end if params
+      params&.each do |key, value|
+        path = path.gsub(":#{key}", value.to_s)
+      end
       path
     end
   end
+  # rubocop:enable Metrics/ClassLength
 end